+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 6.0.0 (April 20, 2022)
+## 6.0.0 (May 13, 2022)
 
 _Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v7.x.x_